Move more non-component code out of BaseHTMLEngineProvider
